Move keyboard navigate mixin into picklist (deprecated)
This removes all mixins from the codebase
